feat: databases one module refactoring
