--- src/dataAccess/databases/mongodb/database.module.ts
@@ -1,15 +1,0 @@
-import { MongooseModule } from '@nestjs/mongoose';
-import { ConfigService } from '@nestjs/config';
-import { Module } from '@nestjs/common';
-
-@Module({
-  imports: [
-    MongooseModule.forRootAsync({
-      inject: [ConfigService],
-      useFactory: (config: ConfigService) => ({
-        uri: config.get<string>('MONGO_URI'),
-      }),
-    }),
-  ],
-})
-export class MongoDatabaseModule {}
