--- src/dataAccess/databases/postgresql/database.module.ts
@@ -1,8 +1,0 @@
-import { databaseProvider } from "./database.service";
-import { Module } from "@nestjs/common";
-
-@Module({
-  imports: [databaseProvider],
-  exports: [databaseProvider],
-})
-export class DatabaseModule {}
